refactor: change dir name from "GameSystem" to  "game-system"
