+++ src/components/game-system/EffectWindow.vue
