+++ src/components/game-system/MediaModal.vue
