+++ src/components/game-system/SideBar.vue
