+++ src/components/game-system/header.vue
