+++ src/components/game-system/hintbutton.vue
